大四_Senior/Visual_Recognition/HW2/src/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
pycocotools>=2.0.6
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
tensorboard
